pyproject.toml
[tool.poetry]
name = "llm-experiments"
version = "0.1.0"
authors = [
    "Ben Jones <btjones.me@gmail.com>"
]
description = "Python Boilerplate contains all the boilerplate you need to create a Python package."
license = "MIT license"
keywords = [
    "FIXME",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]

homepage = "https://github.com/my org/llm-experiments"
repository = "https://github.com/my org/llm-experiments"

readme = "README.md"

[[tool.poetry.source]]
name = "pypi_"
url = "https://pypi.org/simple/"

#[[tool.poetry.source]]
#name = "pypigetfeed"
#url = ""
#secondary = true

# Do not manually edit, use poetry add/remove (--dev) to manage dependencies
[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.7"
tiktoken = "^0.4.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.2"
numpy = "^1.24.3"
langchain = "^0.0.200"
pygame = "^2.5.0"
streamlit = "^1.24.0"
seaborn = "^0.12.2"
scikit-learn = "^1.3.0"
google-cloud-aiplatform = "^1.28.1"
scrubadub = "^2.0.0"
google-api-python-client = "^2.97.0"
google-auth-oauthlib = "^1.0.0"
beautifulsoup4 = "^4.12.2"
selenium = "^4.13.0"
loguru = "^0.7.2"
modin = "^0.25.1"
ray = "^2.8.1"
google-cloud-speech = "^2.23.0"
pydub = "^0.25.1"
plotly = "^5.21.0"
websockets = "^12.0"
streamlit-plotly-events = "^0.0.6"
bigframes = "^1.2.0"

[tool.poetry.dev-dependencies]
assertpy = "*"
bandit = "*"
black = "*"
coverage = "*"
flake8 = "*"
isort = "*"
liccheck = "*"
mypy = "*"
pytest-cov = "*"
pytest-runner = "*"
pytest = "*"
tox-gh-actions = "*"
tox = "*"
toml = "*"
keyring = "*"
artifacts-keyring = "*"
twine = "*"

[tool.pytest.ini_options]
addopts = "--cov=llm_experiments --cov-fail-under=95"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
